NODE-65, test: split runtime tests
