+++ tests/package.json
@@ -4,8 +4,8 @@
   "author": "bifrost-platform",
   "scripts": {
     "test": "mocha -r ts-node/register tests/**/**/*.ts --exit",
-    "test_runtime": "mocha -r ts-node/register runtime/**/**/*runtime*.ts --exit",
-    "test_client": "mocha -r ts-node/register runtime/**/**/*client*.ts --exit",
+    "test_local_runtime": "mocha -r ts-node/register runtime/**/**/*local_runtime*.ts --exit",
+    "test_live_runtime": "mocha -r ts-node/register runtime/**/**/*live_runtime*.ts --exit",
     "run_node": "ts-node commands/run_node.ts",
     "set_node": "ts-node commands/set_node.ts",
     "rewind_node": "ts-node commands/rewind_node.ts",
